Modify playground CLI to use async/await for verify sandboxes extension
