+++ examples/example-app-angular-cli/package.json
@@ -7,6 +7,7 @@
     "ng": "ng",
     "playground:copy": "node ../../scripts/copy.js",
     "playground:run": "node ./node_modules/angular-playground/dist/bin/cli.js",
+    "playground:test": "node ./node_modules/angular-playground/dist/bin/verify-sandboxes/verify-sandboxes.js --path=./src/sandboxes.ts",
     "playground": "npm run playground:copy && npm run playground:run",
     "playground:dev": "npm run build --prefix ../../ && npm run playground",
     "playground2:run": "node ./node_modules/angular-playground/dist/bin/cli.js angular-playground2.json",
